+++ src/cyswllt/__main__.py
@@ -1,0 +1,10 @@
+# Copyright (C) 2026 Chuck Talk, Nordheim Online, LLC <chuck@nordheim.online>
+# This file is part of Cyswllt.
+# Released under the GNU GPL v3 license.
+
+"""Allows ``python -m cyswllt`` to launch the application."""
+
+from cyswllt.main import main
+
+if __name__ == "__main__":
+    main()